Enh/data utils (#92)
* added handling for SSL verification error
* update ASSET and UE tutorials
